commonize field method
