+++ src/bn256/mod.rs
@@ -1,3 +1,4 @@
+mod common;
 mod engine;
 mod fq;
 mod fq12;
